Remove testing on JRuby 9.1 in CI
Tests pass locally, and last commit passed correctly when run
8 months ago (no code changes since). I have many other projects
testing on JRuby 9.1 using minitest-hooks without problems, so
this doesn't appear to be a problem in minitest hooks. I don't
feel debugging this is a good use of my time, so I'm going to
remove jruby-9.1 from the CI matrix.