Make `with_info_handler` take rest args to fix upstream arity change
The upstream `with_info_handler` method was changed to take an extra parameter, which is now breaking the implementation of the method in this gem. This commit fixes the issue by making the method take rest arguments, which will allow it to work with the upstream changes, and any other positional arguments that might be added in the future.